Mike Garratt's security modifications.

+++ home/exim4u/public_html/exim4u/admingroupdelete.php
@@ -4,6 +4,14 @@
   include_once dirname(__FILE__) . "/config/functions.php";
 
   if ($_GET['confirm'] == '1') {
+    # confirm that the user is deleting a group they are permitted to change before going further  
+	$query = "SELECT * FROM groups WHERE id='{$_GET['group_id']}' AND domain_id='{$_SESSION['domain_id']}'";
+	$result = $db->query($query);
+	if ($result->numRows()<1) {
+	  header ("Location: admingroup.php?group_faildeleted={$_GET['localpart']}");
+	  die();  
+	}
+	
     # delete group member first
     $query = "DELETE FROM group_contents WHERE group_id='{$_GET['group_id']}'";
     $result = $db->query($query);
@@ -11,19 +19,17 @@
       # delete group
       $query = "DELETE FROM groups
         WHERE id='{$_GET['group_id']}'
-        AND domain_id={$_SESSION['domain_id']}";
+        AND domain_id='{$_SESSION['domain_id']}'";
       $result = $db->query($query);
       if (!DB::isError($result)) {
         header ("Location: admingroup.php?group_deleted={$_GET['localpart']}");
         die;
       } else {
-        header ("Location: admingroup.php?
-          group_faildeleted={$_GET['localpart']}");
+        header ("Location: admingroup.php?group_faildeleted={$_GET['localpart']}");
         die;
       }
     } else {
-      header ("Location: admingroup.php?
-        group_faildeleted={$_GET['localpart']}");
+      header ("Location: admingroup.php?group_faildeleted={$_GET['localpart']}");
       die;
     }
   } else if ($_GET['confirm'] == 'cancel') {
