+++ home/exim4u/public_html/exim4u/adminuserdelete.php
@@ -4,11 +4,34 @@
   include_once dirname(__FILE__) . '/config/functions.php';
   include_once dirname(__FILE__) . '/config/httpheaders.php';
 
+# confirm that the postmaster is looking to delete a user they are permitted to change before going further  
+$query = "SELECT * FROM users WHERE user_id='{$_GET['user_id']}'
+	AND domain_id='{$_SESSION['domain_id']}'	
+	AND (type='local' OR type='piped')";
+$result = $db->query($query);
+if ($result->numRows()<1) {
+  header ("Location: adminuser.php?faildeleted={$_GET['localpart']}");
+  die();  
+}
+
+  
 if ($_GET['confirm'] == '1') {
+
+  # prevent deleting the last admin
+  $query = "SELECT user_id AS count FROM users 
+    WHERE admin=1 AND domain_id='{$_SESSION['domain_id']}'
+	AND (type='local' OR type='piped')
+    AND user_id!='{$_GET['user_id']}'";
+  $result = $db->query($query);
+  if ($result->numRows() == 0) {
+    header ("Location: adminuser.php?lastadmin={$_GET['localpart']}");
+    die;
+  }  
 
   $query = "DELETE FROM users
     WHERE user_id='{$_GET['user_id']}'
-    AND domain_id={$_SESSION['domain_id']}";
+    AND domain_id='{$_SESSION['domain_id']}'
+	AND (type='local' OR type='piped')";
   $result = $db->query($query);
   if (!DB::isError($result)) {
     $query = "DELETE FROM group_contents WHERE member_id='{$_GET['user_id']}'";
@@ -22,7 +45,8 @@ if ($_GET['confirm'] == '1') {
     die;                                                      
 } else {
   $query = "SELECT user_id AS count FROM users 
-    WHERE admin=1 AND domain_id={$_SESSION['domain_id']}
+    WHERE admin=1 AND domain_id='{$_SESSION['domain_id']}'
+	AND (type='local' OR type='piped')
     AND user_id!='{$_GET['user_id']}'";
   $result = $db->query($query);
   if ($result->numRows() == 0) {
